whitelisting checks: simplify checks by using common base class
